requirements.txt

absl-py==2.3.0
alembic==1.16.1
appdirs==1.4.4
astroid==3.3.9
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
dill==0.3.9
docutils==0.21.2
exceptiongroup==1.2.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.2
grpcio==1.73.0
gurobipy==12.0.2
h5py==3.14.0
humanfriendly==10.0
idna==3.10
immutabledict==4.2.1
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.10
maraboupy==2.0.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multipledispatch==1.0.0
namex==0.1.0
networkx==3.4.2
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnx2pytorch @ git+https://github.com/Verified-Intelligence/onnx2pytorch.git@8447c42c3192dad383e5598edc74dddac5706ee2
onnxconverter-common==1.14.0
onnxoptimizer==0.3.13
onnxruntime==1.22.0
onnxsim==0.4.36
opt_einsum==3.4.0
optree==0.16.0
optuna==4.3.0
ortools==9.12.4544
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250527
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
protobuf==5.29.4
psutil==7.0.0
Pygments==2.19.1
pylint==3.3.6
pyNeVer==0.1.2
pyparsing==3.2.3
PySMT==0.9.6
pytest==8.1.1
pytest-mock==3.14.0
pytest-order==1.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
skl2onnx==1.18.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
statistics==1.0.3.5
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
triton==2.3.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
